Rearrange files to make more organised
