+++ resources/Documentation/installation_guide.md
@@ -2,10 +2,22 @@
 
 **Requirements:**\
 • Python 3.7\
-• Compiler
+• PIP (should be installed with Python)
+• Pygame (Type in Terminal: python3 -m pip install -U pygame --user)\
 
 **To run the game:**
-1. Dowload all files in github repository
-2. Run "main.py"
+1. Dowload all files in github repository\
+2. Run "run_jeff.py"
+
+**Running slow on a Mac?**
+Because of the resolution of the display, Jeff's Lair may run slower on an Apple Mac. 
+You can try this to speed it up:
+
+1. Run the program
+2. In the dock you will see a snake with controllers in his mouth. Right click him.
+3. Go to Options and click “Show in Finder”
+4. Finder will open and you will see a the python application.(Mine was in the shape of rocket with the idle symbol on it.)
+5. Right click the python application and click “Get Info”.
+6. Check the box “Open in Low Resolution”
