+++ resources/LeaderBoard.txt
@@ -1,5 +1,5 @@
 asdaadasdas,94
 adeqwwq qwwq,93
+a,93
+tim,87
 Zhou,80
-Hajar,56
-Geetha,34
